requirements.txt
aiogram
python-dotenv
SQLAlchemy
apscheduler